requirements.txt
streamlit
numpy
pillow
tensorflow
scikit-learn
faiss-cpu
kaggle